Utilisation de fichiers mml pour les tests
