+++ org.xtext.example.mml.tests/mml.py
@@ -1,3 +1,3 @@
 import pandas as pd
-mml_data = pd.read_csv('foo2.csv', sep=';')
+mml_data = pd.read_csv('foo2.csv', sep=',')
 print (mml_data)
